+++ pkg/crud/keyset.go
@@ -1,0 +1,5 @@
+package crud
+
+type hash [25]byte
+
+type keySet map[hash]struct{}